p2p, pot, swarm/pss, swarm/storage, swarm: De-linting, versionfilter
