+++ swarm/pss/pss.go
@@ -215,9 +215,7 @@ func (self *Pss) APIs() []rpc.API {
 			Public:    true,
 		},
 	}
-	for _, auxapi := range self.auxAPIs {
-		apis = append(apis, auxapi)
-	}
+	apis = append(apis, self.auxAPIs...)
 	return apis
 }
 
@@ -388,7 +386,7 @@ func (self *Pss) SetPeerPublicKey(pubkey *ecdsa.PublicKey, topic Topic, address 
 		address: address,
 	}
 	self.pubKeyPoolMu.Lock()
-	if _, ok := self.pubKeyPool[pubkeyid]; ok == false {
+	if _, ok := self.pubKeyPool[pubkeyid]; !ok {
 		self.pubKeyPool[pubkeyid] = make(map[Topic]*pssPeer)
 	}
 	self.pubKeyPool[pubkeyid][topic] = psp
@@ -537,7 +535,7 @@ func (self *Pss) cleanKeys() (count int) {
 					match = true
 				}
 			}
-			if match == false {
+			if !match {
 				expiredtopics = append(expiredtopics, topic)
 			}
 		}
