+++ swarm/pss/pss_go18plus_test.go
@@ -1,4 +1,4 @@
-//-build go17
+// +build go1.8
 package pss
 
 import (
@@ -19,11 +19,13 @@ import (
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/log"
+	"github.com/ethereum/go-ethereum/p2p"
 	"github.com/ethereum/go-ethereum/p2p/discover"
 	"github.com/ethereum/go-ethereum/p2p/simulations"
 	"github.com/ethereum/go-ethereum/p2p/simulations/adapters"
 	"github.com/ethereum/go-ethereum/rpc"
 	"github.com/ethereum/go-ethereum/swarm/network"
+	whisper "github.com/ethereum/go-ethereum/whisper/whisperv5"
 )
 
 // send symmetrically encrypted message between two directly connected peers
@@ -361,7 +363,7 @@ func testNetwork(t *testing.T) {
 				select {
 				case recvmsg := <-msgC:
 					idx, _ := binary.Uvarint(recvmsg.Msg)
-					if recvmsgs[idx] == false {
+					if !recvmsgs[idx] {
 						log.Debug("msg recv", "idx", idx, "id", id)
 						recvmsgs[idx] = true
 						trigger <- id
@@ -542,5 +544,171 @@ func benchmarkAsymKeySend(b *testing.B) {
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		ps.SendAsym(common.ToHex(crypto.FromECDSAPub(&privkey.PublicKey)), topic, msg)
+	}
+}
+
+func BenchmarkSymkeyBruteforceChangeaddr(b *testing.B) {
+	for i := 100; i < 100000; i = i * 10 {
+		for j := 32; j < 10000; j = j * 8 {
+			b.Run(fmt.Sprintf("%d/%d", i, j), benchmarkSymkeyBruteforceChangeaddr)
+		}
+		//b.Run(fmt.Sprintf("%d", i), benchmarkSymkeyBruteforceChangeaddr)
+	}
+}
+
+// decrypt performance using symkey cache, worst case
+// (decrypt key always last in cache)
+func benchmarkSymkeyBruteforceChangeaddr(b *testing.B) {
+	keycountstring := strings.Split(b.Name(), "/")
+	cachesize := int64(0)
+	var ps *Pss
+	if len(keycountstring) < 2 {
+		b.Fatalf("benchmark called without count param")
+	}
+	keycount, err := strconv.ParseInt(keycountstring[1], 10, 0)
+	if err != nil {
+		b.Fatalf("benchmark called with invalid count param '%s': %v", keycountstring[1], err)
+	}
+	if len(keycountstring) == 3 {
+		cachesize, err = strconv.ParseInt(keycountstring[2], 10, 0)
+		if err != nil {
+			b.Fatalf("benchmark called with invalid cachesize '%s': %v", keycountstring[2], err)
+		}
+	}
+	pssmsgs := make([]*PssMsg, 0, keycount)
+	var keyid string
+	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
+	defer cancel()
+	keys, err := wapi.NewKeyPair(ctx)
+	privkey, err := w.GetPrivateKey(keys)
+	if cachesize > 0 {
+		ps = newTestPss(privkey, nil, &PssParams{SymKeyCacheCapacity: int(cachesize)})
+	} else {
+		ps = newTestPss(privkey, nil, nil)
+	}
+	topic := BytesToTopic([]byte("foo"))
+	for i := 0; i < int(keycount); i++ {
+		to := make(PssAddress, 32)
+		copy(to[:], network.RandomAddr().Over())
+		keyid, err = ps.generateSymmetricKey(topic, &to, true)
+		if err != nil {
+			b.Fatalf("cant generate symkey #%d: %v", i, err)
+		}
+		symkey, err := ps.w.GetSymKey(keyid)
+		if err != nil {
+			b.Fatalf("could not retrieve symkey %s: %v", keyid, err)
+		}
+		wparams := &whisper.MessageParams{
+			TTL:      defaultWhisperTTL,
+			KeySym:   symkey,
+			Topic:    whisper.TopicType(topic),
+			WorkTime: defaultWhisperWorkTime,
+			PoW:      defaultWhisperPoW,
+			Payload:  []byte("xyzzy"),
+			Padding:  []byte("1234567890abcdef"),
+		}
+		woutmsg, err := whisper.NewSentMessage(wparams)
+		if err != nil {
+			b.Fatalf("could not create whisper message: %v", err)
+		}
+		env, err := woutmsg.Wrap(wparams)
+		if err != nil {
+			b.Fatalf("could not generate whisper envelope: %v", err)
+		}
+		ps.Register(&topic, func(msg []byte, p *p2p.Peer, asymmetric bool, keyid string) error {
+			return nil
+		})
+		pssmsgs = append(pssmsgs, &PssMsg{
+			To:      to,
+			Payload: env,
+		})
+	}
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		if !ps.process(pssmsgs[len(pssmsgs)-(i%len(pssmsgs))-1]) {
+			b.Fatalf("pss processing failed: %v", err)
+		}
+	}
+}
+
+func BenchmarkSymkeyBruteforceSameaddr(b *testing.B) {
+	for i := 100; i < 100000; i = i * 10 {
+		for j := 32; j < 10000; j = j * 8 {
+			b.Run(fmt.Sprintf("%d/%d", i, j), benchmarkSymkeyBruteforceSameaddr)
+		}
+	}
+}
+
+// decrypt performance using symkey cache, best case
+// (decrypt key always first in cache)
+func benchmarkSymkeyBruteforceSameaddr(b *testing.B) {
+	var keyid string
+	var ps *Pss
+	cachesize := int64(0)
+	keycountstring := strings.Split(b.Name(), "/")
+	if len(keycountstring) < 2 {
+		b.Fatalf("benchmark called without count param")
+	}
+	keycount, err := strconv.ParseInt(keycountstring[1], 10, 0)
+	if err != nil {
+		b.Fatalf("benchmark called with invalid count param '%s': %v", keycountstring[1], err)
+	}
+	if len(keycountstring) == 3 {
+		cachesize, err = strconv.ParseInt(keycountstring[2], 10, 0)
+		if err != nil {
+			b.Fatalf("benchmark called with invalid cachesize '%s': %v", keycountstring[2], err)
+		}
+	}
+	addr := make([]PssAddress, keycount)
+	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
+	defer cancel()
+	keys, err := wapi.NewKeyPair(ctx)
+	privkey, err := w.GetPrivateKey(keys)
+	if cachesize > 0 {
+		ps = newTestPss(privkey, nil, &PssParams{SymKeyCacheCapacity: int(cachesize)})
+	} else {
+		ps = newTestPss(privkey, nil, nil)
+	}
+	topic := BytesToTopic([]byte("foo"))
+	for i := 0; i < int(keycount); i++ {
+		copy(addr[i], network.RandomAddr().Over())
+		keyid, err = ps.generateSymmetricKey(topic, &addr[i], true)
+		if err != nil {
+			b.Fatalf("cant generate symkey #%d: %v", i, err)
+		}
+
+	}
+	symkey, err := ps.w.GetSymKey(keyid)
+	if err != nil {
+		b.Fatalf("could not retrieve symkey %s: %v", keyid, err)
+	}
+	wparams := &whisper.MessageParams{
+		TTL:      defaultWhisperTTL,
+		KeySym:   symkey,
+		Topic:    whisper.TopicType(topic),
+		WorkTime: defaultWhisperWorkTime,
+		PoW:      defaultWhisperPoW,
+		Payload:  []byte("xyzzy"),
+		Padding:  []byte("1234567890abcdef"),
+	}
+	woutmsg, err := whisper.NewSentMessage(wparams)
+	if err != nil {
+		b.Fatalf("could not create whisper message: %v", err)
+	}
+	env, err := woutmsg.Wrap(wparams)
+	if err != nil {
+		b.Fatalf("could not generate whisper envelope: %v", err)
+	}
+	ps.Register(&topic, func(msg []byte, p *p2p.Peer, asymmetric bool, keyid string) error {
+		return nil
+	})
+	pssmsg := &PssMsg{
+		To:      addr[len(addr)-1][:],
+		Payload: env,
+	}
+	for i := 0; i < b.N; i++ {
+		if !ps.process(pssmsg) {
+			b.Fatalf("pss processing failed: %v", err)
+		}
 	}
 }
